+++ ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: tkulket <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/29 00:10:55 by tkulket           #+#    #+#             */
-/*   Updated: 2022/08/29 02:09:01 by tkulket          ###   ########.fr       */
+/*   Updated: 2022/09/06 00:09:43 by tkulket          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int n)
+size_t	ft_strlcpy(char *dst, const char *src, size_t n)
 {
-	unsigned int	size;
-	unsigned int	i;
+	size_t	size;
+	size_t	i;
 
 	size = ft_strlen(src);
 	if (!src && n < size)
