+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: tkulket <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/28 19:44:04 by tkulket           #+#    #+#             */
-/*   Updated: 2022/08/31 08:12:36 by tkulket          ###   ########.fr       */
+/*   Updated: 2022/09/06 16:44:22 by tkulket          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ void	*ft_memchr(const void *s, int c, size_t n);
 int	ft_memcmp(const void *s1, const void *s2, size_t n);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
 void	*ft_memmove(void *dst, const void *src, size_t n);
-void	*ft_memset(void *b, int c, unsigned int n);
+void	*ft_memset(void *b, int c, size_t n);
 void	ft_putchar_fd(char c, int fd);
 void	ft_putendl_fd(char *s, int fd);
 void	ft_putnbr_fd(int n, int fd);
@@ -42,8 +42,8 @@ char	*ft_strchr(const char *s, int c);
 char	*ft_strdup(const char *s1);
 void	ft_striteri(char *s, void (*f)(unsigned int, char*));
 char	*ft_strjoin(char const *s1, char const *s2);
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int n);
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int n);
+size_t	ft_strlcat(char *dst, const char *src, size_t n);
+size_t	ft_strlcpy(char *dst, const char *src, size_t n);
 size_t	ft_strlen(const char *str);
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int	ft_strncmp(const char *s1, const char *s2, unsigned int n);
